+++ agent/CHANGELOG.md
@@ -2,6 +2,9 @@
 - 文本暂存 tool，允许模型主动调用，封存上下文中的指定数据到内存，只在上下文中保留id和缩略描述
 
 ## LOG
+- 2026/06/02
+    - 添加 middleware，拦截 `agentCall`、`modelCall`、`toolCall`
+    - 添加 skill 解析
 - 2026/05/22
     - `toolcall`:
         - 支持`filesystem`、`websearch`、`execute_command`、`string_util`